+++ 0.56.b/graysvr/CClientMsg.cpp
@@ -591,7 +591,7 @@ int CClient::addContents( const CItemContainer * pContainer, bool fCorpseEquip, 
 	{
 		if ( count >= MAX_ITEMS_CONTENT )
 		{
-			g_Log.EventWarn("Too many items in container '%s' uid=0%x items=%d\n", pContainer->GetName(), pContainer->GetUID(), pContainer->GetCount());
+			g_Log.EventWarn("Too many items in container '%s' uid=0%x items=%d\n", pContainer->GetName(), (DWORD)pContainer->GetUID(), pContainer->GetCount());
 			break;
 		}
