+++ 0.56.b/graysvr/CSector.cpp
@@ -1052,7 +1052,7 @@ void CSector::OnTick(int iPulseCount)
 
 		EXC_DEBUGSUB_START;
 		CPointMap pt = GetBasePoint();
-		g_Log.EventDebug("char 0%lx '%s'\n", pChar->GetUID(), pChar->GetName());
+		g_Log.EventDebug("char 0%lx '%s'\n", (DWORD)pChar->GetUID(), pChar->GetName());
 		g_Log.EventDebug("sector #%d [%d,%d,%d,%d]\n", GetIndex(),  pt.m_x, pt.m_y, pt.m_z, pt.m_map);
 		EXC_DEBUGSUB_END;
 	}
@@ -1094,7 +1094,7 @@ void CSector::OnTick(int iPulseCount)
 
 		EXC_DEBUGSUB_START;
 		CPointMap pt = GetBasePoint();
-		g_Log.EventError("item 0%lx '%s' [timer=%d, type=%d]\n", pItem->GetUID(), pItem->GetName(), pItem->GetTimerAdjusted(), (int)pItem->GetType());
+		g_Log.EventError("item 0%lx '%s' [timer=%d, type=%d]\n", (DWORD)pItem->GetUID(), pItem->GetName(), pItem->GetTimerAdjusted(), (int)pItem->GetType());
 		g_Log.EventError("sector #%d [%d,%d,%d,%d]\n", GetIndex(),  pt.m_x, pt.m_y, pt.m_z, pt.m_map);
 		
 		EXC_DEBUGSUB_END;
@@ -1103,14 +1103,14 @@ void CSector::OnTick(int iPulseCount)
 		catch ( CGrayError &e )
 		{
 			CPointMap pt = GetBasePoint();
-			g_Log.EventError("CGrayError: item 0%lx '%s' [timer=%d, type=%d]\n", pItem->GetUID(), pItem->GetName(), pItem->GetTimerAdjusted(), (int)pItem->GetType());
+			g_Log.EventError("CGrayError: item 0%lx '%s' [timer=%d, type=%d]\n", (DWORD)pItem->GetUID(), pItem->GetName(), pItem->GetTimerAdjusted(), (int)pItem->GetType());
 			g_Log.EventError("sector #%d [%d,%d,%d,%d]\n", GetIndex(),  pt.m_x, pt.m_y, pt.m_z, pt.m_map);
 			EXC_CATCH_SUB(&e, "Sector");
 		}
 		catch (...)
 		{
 			CPointMap pt = GetBasePoint();
-			g_Log.EventError("...: item 0%lx '%s' [timer=%d, type=%d]\n", pItem->GetUID(), pItem->GetName(), pItem->GetTimerAdjusted(), (int)pItem->GetType());\
+			g_Log.EventError("...: item 0%lx '%s' [timer=%d, type=%d]\n", (DWORD)pItem->GetUID(), pItem->GetName(), pItem->GetTimerAdjusted(), (int)pItem->GetType());\
 			g_Log.EventError("sector #%d [%d,%d,%d,%d]\n", GetIndex(),  pt.m_x, pt.m_y, pt.m_z, pt.m_map);
 			EXC_CATCH_SUB(NULL, "Sector");
 		}
